Added Sync endpoint separate from objective and element retrieval. (#14)
* Use datetime2 instead of datetime. Fix descriptors type for PerformanceEvaluation.

* Add method to retrieve PerformanceEvaluations

* Sync PerformanceEvaluations on evaluation controller.

* Fix descriptors types in PerformanceEvaluation. Add explicit cast to SDK model.

* Retrieve PerformanceEvaluations from local DB instead of EdFi API.

* Match EducationOrganizationId column type.

* Add method to retrieve PerformanceEvaluations.

* Create method to Sync dependencies.
Filter evaluationObjectives by performanceEvaluation.

* Change GetPerformanceEvaluationById signature.
Change GetEvaluationObjectivesElementsTitles parameter routing.

+++ src/webapi/Evaluations/Models/PerformanceEvaluation.cs
@@ -12,7 +12,7 @@ namespace eppeta.webapi.Evaluations.Models;
 [Table("PerformanceEvaluation", Schema = "eppeta")]
 public partial class PerformanceEvaluation
 {
-    public int EducationOrganizationId { get; set; }
+    public long EducationOrganizationId { get; set; }
 
     public string EvaluationPeriodDescriptor { get; set; }
